아이 키, 몸무게 실수형 변경

+++ src/main/java/com/choongang/frombirth_backend/model/dto/ChildrenDTO.java
@@ -25,9 +25,9 @@ public class ChildrenDTO {
     private String bloodType;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Integer birthWeight;
+    private Double birthWeight;
     private LocalTime birthTime;
     private String profilePicture;
-    private Integer birthHeight;
+    private Double birthHeight;
 }
